Remove type casts from allocation macros.
As the removed comment notes, it was sometimes necessary long-ago to
cast allocations. This is no longer the case, but nvi2's allocation
macros still take type parameters. We've already removed them in
OpenBSD's vi.

This commit also removes a needless NULL-check before free() in
REALLOC(). POSIX specifies that free() is NULL-safe, and all modern
platforms conform to this.