Fix -c option in wide char mode
When ex_run_str is called with nocopy=1 in wide char mode, it causes the command pointer to refer to the internal conv buffer from the call to CHAR2INT prior. This buffer is then corrupted later on with line data by db_get in the call to FILE2INT there. To test: echo foobar > file && nvi -e -c '/foo/p' file.