Add Vim-style expandtab option.
The expandtab setting expands tabs to spaces in insert mode as well
as when shifting and indenting/outdenting.  This is very useful
when working on a code-base where the style dictates using spaces
instead of tabs for indentation.  Originally ported from NetBSD,
but has been extended to behave more like the vim implementation.