vi.1: Link macro for hyperlink (FreeBSD #272940)
The trailing full-stop in the hyperlink to the Vi/Ex Reference Manual in the see also section breaks it in some implementations, so replace that structure with the mdoc(7) macro designed for hyperlinks. The title of the paper seems capitalized in the paper itself so do that here as well. This fixes FreeBSD bug #272940.

While here, a little cleaning:
- bump date
- shuffle 7 sentences onto their own lines
- prepend a zero width space to the ! in :q!
- replace the a few left brackets with a roff(7) sequence to display them
- Follow mdoc(7) convention for the name section by breaking the name macros onto their own lines.