Generate options_def.h, ex_def.h, and version.h in build directory
These three headers were generated in the source directories, which was
not a good idea since we're using an out-of-source-tree building scheme
now.  It also caused difficulties in importing this software to other
projects.

So similar to the previous commit, generate these headers in the build
directory.  Now the source tree is no longer modified during the build.