Make counted p more usable (#150)
Contributed by: Walter Alejandro Iglesias <wai@roquesor.com>

An implementation detail of `p` with a count made it not as usable as
`P`. Users have to switch to `P` with a count to get sensible content.
The patch solves the issue by factoring the cursor placement out of
`p`/`P` with a count. After the change, `p` also collects content
without interleaving.

Closes: #146