Simplify options initialization (#152)
After #127, the `OI` macro has awkward semantics. Previously, the
expanded macro always parsed setting out of a known buffer, but it lets
the caller specify an input. If the input happens to be the known
buffer, the expanded macro avoids undefined behavior caused by `memcpy`
between memory locations that overlap. Observing that memory locations
with different content won't start at the same address, `strcmp` is a
weird way to meet that purpose.

The proposed `OI` is variadic and performs `snprintf` internally.
Double-checked that we don't have a dynamic source of string template.

---------

Co-authored-by: turistu <turistu@gmail.com>