Insert empty line for trailing delimiter or newline (#157)
A trailing delimiter should cause an empty line to be inserted and this
is also what the version of vi that comes with AIX and that identifies
itself as "Version 3.10" does.

Keep track of the removal of a trailing delimiter or newline so that an
empty line can be inserted when this removal results in an empty
trailing text.

Closes: #135

Co-authored-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>