Handle out-of-band lines in expandtab (#154)
The procedure that causes the bug:
* open vi with set expandtab
* yy -> yank something
* :di b -> shows that the default buffer have the text copied
* :!ls >/dev/null run an external command without filtering lines, all
commands are the same
* :di b -> already shows something strange, outputting: **********
default buffer (character mode) note that when the buffers are empty :di
b shows: No cut buffers to display.
* p -> trying to paste segfault

A bug that does not segfault occurs also with the bang command, but when
filtering lines:
* open vi with set expandtab
* yy -> yank something and then move to an other line
* :di b -> shows that the default buffer have the text copied
* :.!ls >/dev/null run an external command filtering any range of lines,
both one address or two is the same, noticing the content of the line
you are on
* :di b -> shows that the line(s) being put into the default buffer
is(are) the line(s) that is(are) now under the cursor, not the original
deleted ones

Running the bang without filtering files leaves the sp->gp->dcb_store as
a dangling pointer.

Closes: #151

Co-authored-by: Jerry Fletcher <mail@jerryfletcher.org>